+++ src/qp_vault/__init__.py
@@ -26,7 +26,7 @@ Quick start:
 Docs: https://github.com/quantumpipes/vault
 """
 
-__version__ = "1.5.1"
+__version__ = "1.5.2"
 __author__ = "Quantum Pipes Technologies, LLC"
 __license__ = "Apache-2.0"